css layout task10.css added
This is CSS for task10.html. These files contain a body style whose height is 100vh declared then align items in the center, then the container will display as a flex. The big-box width and height is 550px and 500px simultaneously border will be 1px solid gray color position will be relative 'dot are in display: inline-block' dot's color will be 'red, blue and green' big-container display flex it's width is 400px and height is 200px finally gap is 80px.  small-flex div will also display flex it's background color is black. Then small-box position will be absolute it's width and hight is 100px.